requirements.txt
pandas
streamlit
plotly.express